Extract Editor behaviour to separate js file
