+++ lib/repl.js
@@ -1,10 +1,7 @@
 'use babel';
 
 const Status = require('./status')
-
-const EVAL_LINE = 'line'
-const EVAL_MULTI_LINE = 'multi_line'
-const EVAL_WHOLE_EDITOR = 'whole_editor'
+const Editor = require('./editor')
 
 let status = new Status()
 
@@ -25,7 +22,7 @@ export default class REPL {
 
     atom.commands.add('atom-workspace', {
       'tidalcycles:boot': () => {
-        if (this.editorIsTidal()) {
+        if (Editor.isTidal()) {
           this.start();
         } else {
           console.error('Not a .tidal file.');
@@ -38,19 +35,13 @@ export default class REPL {
     });
 
     atom.commands.add('atom-text-editor', {
-      'tidalcycles:eval': () => this.eval(EVAL_LINE, false),
-      'tidalcycles:eval-multi-line': () => this.eval(EVAL_MULTI_LINE, false),
-      'tidalcycles:eval-whole-editor': () => this.eval(EVAL_WHOLE_EDITOR, false),
-      'tidalcycles:eval-copy': () => this.eval(EVAL_LINE, true),
-      'tidalcycles:eval-multi-line-copy': () => this.eval(EVAL_MULTI_LINE, true),
+      'tidalcycles:eval': () => this.eval(Editor.LINE, false),
+      'tidalcycles:eval-multi-line': () => this.eval(Editor.MULTI_LINE, false),
+      'tidalcycles:eval-whole-editor': () => this.eval(Editor.WHOLE_EDITOR, false),
+      'tidalcycles:eval-copy': () => this.eval(Editor.LINE, true),
+      'tidalcycles:eval-multi-line-copy': () => this.eval(Editor.MULTI_LINE, true),
       'tidalcycles:hush': () => this.hush()
     });
-  }
-
-  editorIsTidal() {
-    var editor = this.getEditor();
-    if (!editor) return false;
-    return editor.getGrammar().scopeName === 'source.tidalcycles';
   }
 
   start() {
@@ -141,16 +132,12 @@ export default class REPL {
     this.ghci.writeLine(command);
   }
 
-  getEditor() {
-    return atom.workspace.getActiveTextEditor();
-  }
-
   eval(evalType, copy) {
-    if (!this.editorIsTidal()) return;
+    if (!Editor.isTidal()) return;
 
     if (!this.ghci) this.start();
 
-    let evaluations = this.currentEvaluations(evalType);
+    let evaluations = Editor.currentEvaluations(evalType);
     for (var i = 0; i < evaluations.length; i++) {
       let evaluation = evaluations[i]
       var expression = evaluation.expression;
@@ -164,9 +151,9 @@ export default class REPL {
     var self = this;
     if (!expression || !range) return;
 
-    var unflash = self.evalFlash(range);
+    var unflash = Editor.evalFlash(range);
     if (copy) {
-      self.copyRange(range);
+      Editor.copyRange(range);
     }
 
     self.tidalSendExpression(expression);
@@ -182,140 +169,4 @@ export default class REPL {
     }
   }
 
-  currentEvaluations(evalType) {
-    var editor = this.getEditor();
-    if (!editor) return;
-
-    var selection = editor.getLastSelection();
-    var expression = selection.getText();
-
-    if (expression) {
-      var range = selection.getBufferRange();
-      return [{ expression, range }];
-    } else {
-      switch (evalType) {
-        case EVAL_LINE:
-          return this.getLineExpression(editor);
-        case EVAL_MULTI_LINE:
-          return this.getMultiLineExpression(editor);
-        case EVAL_WHOLE_EDITOR:
-          return this.getWholeEditorExpressions(editor);
-      }
-    }
-  }
-
-  copyRange(range) {
-    var editor = this.getEditor();
-    var endRow = range.end.row;
-    endRow++
-    var text = editor.getTextInBufferRange(range);
-    text = '\n' + text + '\n';
-
-    if (endRow > editor.getLastBufferRow()) {
-      text = '\n' + text
-    }
-
-    editor.getBuffer().insert([endRow, 0], text);
-  }
-
-  getLineExpression(editor) {
-    var cursor = editor.getCursors()[0];
-    var range = cursor.getCurrentLineBufferRange();
-    var expression = range && editor.getTextInBufferRange(range);
-    return [{ expression, range }];
-  }
-
-  getMultiLineExpression(editor) {
-    var range = this.getCurrentParagraphIncludingComments(editor);
-    var expression = editor.getTextInBufferRange(range);
-    return [{ expression, range }];
-  }
-
-  getWholeEditorExpressions(editor) {
-    let wholeEditor = {
-      start: { row: 0, column: 0 },
-      end: { row: editor.getLineCount(), column: 0 }
-    }
-    let expression = editor.getTextInBufferRange(wholeEditor);
-
-    return expression.split('\n')
-      .reduce((blocks, item, index) => {
-        if (blocks.length === 0) {
-          blocks.push({ rows: [] })
-        }
-
-        if (item.trim().length === 0) {
-          blocks.push({ rows: [] })
-        } else {
-          let block = blocks[blocks.length - 1]
-          if (!block.start) {
-            block.start = index
-          }
-          block.rows.push(item)
-          blocks[blocks.length - 1] = block
-        }
-
-        return blocks
-      }, [])
-      .filter(block => block.rows.length > 0)
-      .map(block => {
-        return {
-          expression: block.rows.join('\n'),
-          range: {
-            start: { row: block.start, column: 0},
-            end: { row: block.start + block.rows.length, column: 0},
-          }
-        }
-      })
-  }
-
-  getCurrentParagraphIncludingComments(editor) {
-    var cursor = editor.getLastCursor();
-    var startRow = endRow = cursor.getBufferRow();
-    var lineCount = editor.getLineCount();
-
-    // lines must include non-whitespace characters
-    // and not be outside editor bounds
-    while (/\S/.test(editor.lineTextForBufferRow(startRow)) && startRow >= 0) {
-      startRow--;
-    }
-    while (/\S/.test(editor.lineTextForBufferRow(endRow)) && endRow < lineCount) {
-      endRow++;
-    }
-    return {
-      start: {
-        row: startRow + 1,
-        column: 0
-      },
-      end: {
-        row: endRow,
-        column: 0
-      },
-    };
-  }
-
-  evalFlash(range) {
-    var editor = this.getEditor();
-    var marker = editor.markBufferRange(range, {
-      invalidate: 'touch'
-    });
-
-    var decoration = editor.decorateMarker(
-      marker, {
-        type: 'line',
-        class: 'eval-flash'
-      });
-
-    // return fn to flash error / success and destroy the flash
-    return function(cssClass) {
-      decoration.setProperties({
-        type: 'line',
-        class: cssClass
-      });
-      var destroy = function() {
-        marker.destroy();
-      };
-      setTimeout(destroy, 120);
-    };
-  }
 }
